% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minModuleEx-package.R
\docType{package}
\name{minModuleEx}
\alias{minModuleEx}
\alias{minModuleEx-package}
\title{Minimal Rcpp Module Example}
\description{
minModuleEx
}
\details{
A minimal example of using Rcpp modules in an Rcpp package
}
\author{
James Curran
}

